Streamlined solver spawning, added env config on express-app
