+++ apps/express-app/.env.example
@@ -1,7 +1,11 @@
-#NX_AOC_URL="https://adventofcode.com"
-
 # Replace this with your own sessionid from your adventofcode login.
 NX_AOC_SESSION="SESSION"
 
 # Replace this with your MongoDB URL (see base README for guidance)
 NX_MONGO_URL="MONGO_URL"
+
+#The following are optional env variables intended for express-app development
+#NX_AOC_URL="https://adventofcode.com"
+#NX_SOLVER_URL="http://localhost"
+#NX_SOLVER_PORT=3000
+#NX_SOLVER_DEBUG_PORT=19229
