+++ apps/express-app/project.json
@@ -35,7 +35,8 @@
           "buildTarget": "express-app:build:development"
         },
         "production": {
-          "buildTarget": "express-app:build:production"
+          "buildTarget": "express-app:build:production",
+          "watch": false
         }
       }
     },
