+++ libs/express/controller/solver/src/lib/solverController.ts
@@ -5,31 +5,41 @@ import { stderr, stdin, stdout } from 'process';
 
 import { InputModel } from '@aon/util-types';
 
-const solver = spawn('yarn', ['nx', 'run', 'solver-app:serve:develop']);
+const { NX_SOLVER_URL, NX_SOLVER_PORT = 3000, NX_SOLVER_DEBUG_PORT = 19229 } = process.env;
+const baseUrl = `${NX_SOLVER_URL || 'http://localhost'}:${NX_SOLVER_PORT}`;
+const solveUrl = new URL('/solve', baseUrl);
+const cancelUrl = new URL('/cancel', baseUrl);
 
-stdin.on('data', data => solver.stdin.write(data));
-solver.stdout.on('data', data => stdout.write(data));
-solver.stderr.on('data', data => stderr.write(data));
+if (!NX_SOLVER_URL) {
+  // If the solver URL was manually set, presumably it's being run elsewhere and we don't want to spawn it as a child process.
 
-const url = '';
+  const solver = spawn('yarn', ['nx', 'run', 'solver-app:serve:develop', `--port=${NX_SOLVER_DEBUG_PORT}`]);
+
+  stdin.on('data', data => solver.stdin.write(data));
+  solver.stdout.on('data', data => stdout.write(data));
+  solver.stderr.on('data', data => stderr.write(data));
+}
+
 export const postSolution = async (req: Request, res: Response) => {
   const { id, part } = req.body;
 
   if (id && part) {
-    try {
-      const { year, day, input } = await InputModel.findById(id).select(['year', 'day', 'input']).exec();
-      if (year && day && input) {
-        const result = await fetch(url, { method: 'POST', body: JSON.stringify({ year, day, input, part }) }).then(response =>
-          response.json(),
-        );
-        res.status(201).json(result);
-      } else {
-        throw new Error('Input not found');
-      }
-    } catch (err) {
-      res.status(400).json({ message: err.message, time: 0 });
-    }
+    InputModel.findById(id)
+      .select(['year', 'day', 'input'])
+      .exec()
+      .then(({ year, day, input }) =>
+        fetch(solveUrl, {
+          method: 'POST',
+          headers: { 'Content-type': 'application/json' },
+          body: JSON.stringify({ year, day, input, part }),
+        }),
+      )
+      .then(response => response.json())
+      .then(result => res.status(201).json(result))
+      .catch(err => res.status(500).json({ message: err.message }));
   } else {
-    fetch(url, { method: 'POST' });
+    fetch(cancelUrl, { method: 'POST' })
+      .then(response => res.sendStatus(response.status))
+      .catch(err => res.status(500).json({ message: err.message }));
   }
 };
